% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quiz_module.R
\name{frame_question}
\alias{frame_question}
\title{Create a question with random distractors}
\usage{
frame_question(
  items,
  ndistractor = 5,
  direction = "forward",
  k = sample(length(items), 1)
)
}
\description{
Create a question with random distractors
}
