% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradescopeQ.R
\name{gradescopeQ}
\alias{gradescopeQ}
\title{Render multiple choice questions for GradeScope online assignment}
\usage{
gradescopeQ(
  prompt = "The question prompt",
  ...,
  id = NULL,
  right_one = NULL,
  inline = FALSE,
  random_answer_order = FALSE,
  allow_retry = TRUE,
  correct = "Right!",
  incorrect = "Sorry.",
  message = NULL,
  post_message = NULL,
  submit_button = "Check answer",
  try_again_button = "Try again",
  allow_multiple_correct = FALSE
)
}
\description{
Render multiple choice questions for GradeScope online assignment
}
\examples{
foo <- gradescopeQ("What is $$\\\\int x^2 dx \\\\mbox{?}$$",
           "+an indefinite integral+" = "That's all I wanted.",
           "Something I don't understand" = "Keep studying",
           "a definite integral" = "No, there are no limits of integration as in $\\\\int_a^b$."
           )
}
