% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etudeQ.r
\name{master_id}
\alias{master_id}
\title{Set a master ID for the whole document
when questions are created with etude2 they will
get assigned a unique ID within the document}
\usage{
master_id(id)
}
\description{
Set a master ID for the whole document
when questions are created with etude2 they will
get assigned a unique ID within the document
}
