% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etudeQ.r
\name{static_knitr_format}
\alias{static_knitr_format}
\title{Figure out if the function is being called from knitr with runtime static
(that is, not a learnr document)}
\usage{
static_knitr_format()
}
\description{
Figure out if the function is being called from knitr with runtime static
(that is, not a learnr document)
}
