% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradescopeQ.R
\name{gradescopeQ}
\alias{gradescopeQ}
\title{Render multiple choice questions for GradeScope online assignment}
\usage{
gradescopeQ(
  prompt = "The question prompt",
  ...,
  id = NULL,
  right_one = NULL,
  inline = FALSE,
  random_answer_order = FALSE,
  allow_retry = TRUE,
  correct = "Right!",
  incorrect = "Sorry.",
  message = NULL,
  post_message = NULL,
  submit_button = "Check answer",
  try_again_button = "Try again",
  allow_multiple_correct = FALSE
)
}
\description{
Render multiple choice questions for GradeScope online assignment
}
